web/backend/with_cash/app/requirements.txt
fastapi==0.115.0 uvicorn==0.30.6 
sqlalchemy==2.0.35 
psycopg2-binary==2.9.9 
pydantic-settings==2.5.2 
python-jose==3.3.0 
bcrypt==4.2.0 
redis==5.0.8 
stripe==10.12.0 
google-auth==2.35.0